[SVN] r3169 twister-lib-android in /Libraries/twister-lib-android/ [ADD] Test for missing rotated image after taking picture [FIX] Only clear local image state when creating the activity the first time

+++ libs/twister-lib-android/capture_image/src/main/java/net/twisterrob/android/activity/CaptureImage.java
@@ -114,9 +114,11 @@ public class CaptureImage extends Activity implements ActivityCompat.OnRequestPe
 			return;
 		} else {
 			mTargetFile = new File(output);
-			LOG.trace("Clear image at {}", mTargetFile);
-			//noinspection ResultOfMethodCallIgnored best effort, try to prevent leaking old image
-			mTargetFile.delete();
+			if (savedInstanceState == null) {
+				LOG.trace("Clear image at {}", mTargetFile);
+				//noinspection ResultOfMethodCallIgnored best effort, try to prevent leaking old image
+				mTargetFile.delete();
+			}
 		}
 
 		setContentView(R.layout.activity_camera);
